#134 basic user's number is integer, but 100 quid plan is double. read accordingly

+++ bouncer/src/repo/core/model/bson/repo_bson_user.h
@@ -242,7 +242,7 @@ namespace repo {
 				* Get the current quota of the user
 				* @return returns the quota available to the user
 				*/
-				uint64_t getQuota() const;
+				double getQuota() const;
 
 			private:
 				/**
